convert library resources to proper custom resources
library cookbooks are not a recommended cookbook pattern

see:

https://docs.chef.io/custom_resources_notes.html

and issues like this which are directly related to trying to
fight the resources DSL:

https://github.com/chef-cookbooks/httpd/issues/97
